+++ RaymondMaarloeve/Assets/Skybox/CubeMap/Materials/Sky_LowPoly_01_Day_a.mat
@@ -65,7 +65,7 @@ Material:
     - _Cutoff: 0.5
     - _DstBlend: 0
     - _EnvironmentReflections: 1
-    - _Exposure: 0.94195944
+    - _Exposure: 0.29400054
     - _GlossMapScale: 0
     - _Glossiness: 0
     - _GlossyReflections: 0
@@ -77,7 +77,7 @@ Material:
     - _OcclusionStrength: 1
     - _QueueOffset: 0
     - _ReceiveShadows: 1
-    - _Rotation: 36.401764
+    - _Rotation: 20.727497
     - _Smoothness: 0.5
     - _SmoothnessTextureChannel: 0
     - _SpecularHighlights: 1
